NVENC AV1 can only do 8-bit 4:2:0
NVENC doesn't support 4:2:2 at all.  It does support 4:4:$, but not
for AV1 currently.

NVENC doesn't support 12-bit at all, I think.  It does support 10-bit,
including for AV1, but when I do it the movie is bad, doesn't seem
to have anything to do with the source movie.  Feels like an Nvidia
bug.