NVENC 10-bit is back
It was my bug, not Nvidia's.  The YUV 10-bit buffer uses 16-bit
pixels but expects the 10-bit data to be in the high bits, so
I'll treat them like full 16-bit pixels.

If a user specifies NVENC and tries to do 12-bit, I'll just
do 10-bit instead so I don't have to worry about taking away
the 12-bit option in the UI.