Add Rec. 601/709 control, support.
There are a lot more color spaces out there (Rec. 2020 for example), but
we're just going to have Rec. 601 and 709.  Previously I was assuming
Rec. 601 for everything because that's what WebM seemed to do and what
various readers were doing.  At some point color metadata was added to
the container, but I'm not sure how many people are using it.  But
mainly these days it seemed a little ridiculous to default to Rec. 601
when Rec. 709 (the HDTV standard) is much more widely used.  In addition
to tagging the file, this most impacts RGB to YCrCb conversion where the
two use different matrices.

Could add Rec. 2020 and others in the future, although that is a
different beast because it uses different color primaries.  Not clear to
me right now how to get Premiere to do the conversion or how Premiere's
rudimentary color management works.